- Updating the ratios to actually reflect doubles  - Updated player controller for a less redundant api

+++ src/main/java/com/crow/blood/on/the/clocktower/model/statistics/Ratio.java
@@ -32,7 +32,7 @@ public class Ratio {
 
 	@JsonIgnore
 	public double getRatio() {
-		return positive / getTotal();
+		return ((double) positive) / getTotal();
 	}
 
 	public long getTotal() {
